% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{bf_to_pvalue}
\alias{bf_to_pvalue}
\title{Convert Bayes factor to approximate p-value}
\usage{
bf_to_pvalue(bf)
}
\arguments{
\item{bf}{Bayes factor (BF10)}
}
\value{
Approximate p-value
}
\description{
Convert Bayes factor to approximate p-value
}
\keyword{internal}
