% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{build_pairwise_results}
\alias{build_pairwise_results}
\title{Build results tibble from pairwise tests}
\usage{
build_pairwise_results(
  test_results,
  data,
  compare,
  ref,
  test,
  alpha,
  fdr_method
)
}
\arguments{
\item{test_results}{List of test results for each peptide}

\item{data}{Original pepdiff_data object}

\item{compare}{Factor being compared}

\item{ref}{Reference level}

\item{test}{Test method name}

\item{alpha}{Significance threshold}

\item{fdr_method}{FDR correction method}
}
\value{
A tibble in long format
}
\description{
Build results tibble from pairwise tests
}
\keyword{internal}
