% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{build_replicate_matrix}
\alias{build_replicate_matrix}
\title{Build replicate matrix from pepdiff_data}
\usage{
build_replicate_matrix(data, compare, level)
}
\arguments{
\item{data}{pepdiff_data object}

\item{compare}{Factor column name}

\item{level}{Level of the factor to extract}
}
\value{
Matrix with peptides as rows and replicates as columns
}
\description{
Pivots data to matrix format (peptides × replicates) for a given factor level.
}
\keyword{internal}
