% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{build_results_tibble}
\alias{build_results_tibble}
\title{Build results tibble from model output}
\usage{
build_results_tibble(model_results, data, compare, method, alpha, fdr_method)
}
\arguments{
\item{model_results}{List of results from run_models()}

\item{data}{Original pepdiff_data object}

\item{compare}{Factor being compared}

\item{method}{Method used}

\item{alpha}{Significance threshold}

\item{fdr_method}{FDR correction method}
}
\value{
A tibble in long format
}
\description{
Combines model results into the standard long-format results tibble.
}
\keyword{internal}
