% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepdiff.R
\name{calc_de}
\alias{calc_de}
\title{calculate differential}
\usage{
calc_de(
  df,
  iters = 1000,
  treatment = NA,
  t_seconds = NA,
  control = NA,
  c_seconds = NA
)
}
\description{
calculate differential
}
