% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{calc_rank_product}
\alias{calc_rank_product}
\title{Calculate rank products for two groups}
\usage{
calc_rank_product(control, treatment)
}
\arguments{
\item{control}{Control group values}

\item{treatment}{Treatment group values}
}
\value{
List with rp_up and rp_down
}
\description{
For a single peptide, calculates the rank product statistic based on
all pairwise fold changes between treatment and control replicates.
}
\keyword{internal}
