% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{calc_t_statistic}
\alias{calc_t_statistic}
\title{Calculate t-statistic for two groups}
\usage{
calc_t_statistic(x, y)
}
\arguments{
\item{x}{First group values}

\item{y}{Second group values}
}
\value{
The t-statistic
}
\description{
Calculate t-statistic for two groups
}
\keyword{internal}
