% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_zeros}
\alias{check_zeros}
\title{Check for zeros in values (used before GLM fitting)}
\usage{
check_zeros(values)
}
\arguments{
\item{values}{Numeric vector}
}
\value{
TRUE invisibly if no zeros/negatives, otherwise throws error
}
\description{
Check for zeros in values (used before GLM fitting)
}
\keyword{internal}
