% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepdiff.R
\name{combine_tech_reps}
\alias{combine_tech_reps}
\title{combine tech replicates into one biological replicate measurement}
\usage{
combine_tech_reps(df)
}
\arguments{
\item{df}{dataframe; typically from `import_data()`}
}
\value{
dataframe
}
\description{
combine tech replicates into one biological replicate measurement
}
