% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{compare_calls}
\alias{compare_calls}
\title{compare sets of significant peptides called by the used data}
\usage{
compare_calls(r, sig = 0.05)
}
\arguments{
\item{r}{result dataframe typically from `compare()`}

\item{sig}{significance cut-off to select peptides}
}
\value{
UpSet plot
}
\description{
produces an UpSet plot showing intersections and set-size of the different
sets of significant peptides called by the methods used in the provided result
dataframe
}
