% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepdiff.R
\name{compare_many}
\alias{compare_many}
\title{Compare peptide quantities for multiple experimental conditions using selected statistical tests.}
\usage{
compare_many(
  df,
  comparison,
  iters = 1000,
  tests = c("bootstrap_t"),
  log = FALSE,
  base = 2
)
}
\arguments{
\item{df}{A dataframe containing peptide quantification data, typically obtained from `import_data()`.}

\item{comparison}{A comparison dataframe specifying the experimental conditions to compare. Each row defines a pairwise condition comparison, including treatment, control, and time points. If not already a data frame, you can provide a path to a CSV file, and it will be read into a data frame.}

\item{iters}{The number of iterations to perform for iterative tests. Default is 1000.}

\item{tests}{#' @param tests character vector of tests to use, one or more of: `norm_quantile`, `bootstrap_t`, `wilcoxon`, `kruskal-wallis`, `rank_product`, `gamma`, `eb`}

\item{log}{should the data be logged prior to analysis}

\item{base}{base to use for logging (2)}
}
\value{
A list of dataframes, where each dataframe represents the comparison results for a pairwise condition comparison. The names of the list elements are constructed based on the conditions being compared.
}
\description{
This function compares peptide quantities between multiple experimental conditions specified in a comparison dataframe. It applies selected statistical tests to identify differentially abundant peptides for each pairwise condition comparison.
}
\examples{
\dontrun{
# Example using imported data and a comparison dataframe:
data_file <- "peptide_quantification_data.csv"
comparison_data <- "comparison_conditions.csv"
imported_data <- import_data(data_file)
comparison_result <- compare_many(imported_data, comparison_data,
tests = c("bootstrap_t", "wilcoxon"))
}
}
\seealso{
Other statistical analysis: 
\code{\link{compare}()},
\code{\link{get_bootstrap_percentile}()},
\code{\link{get_kruskal_percentile}()},
\code{\link{get_percentile_lowest_observed_value_iterative}()},
\code{\link{get_rp_percentile}()},
\code{\link{get_wilcoxon_percentile}()}
}
\concept{statistical analysis}
\keyword{data}
