% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_pairwise}
\alias{compare_pairwise}
\title{Compare using pairwise tests}
\usage{
compare_pairwise(data, compare, ref, within, test, alpha, fdr_method)
}
\description{
Compare using pairwise tests
}
\keyword{internal}
