% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_within_strata}
\alias{compare_within_strata}
\title{Compare within strata}
\usage{
compare_within_strata(
  data,
  compare,
  ref,
  within,
  method,
  test,
  alpha,
  fdr_method,
  bf_threshold = 3
)
}
\arguments{
\item{data}{pepdiff_data object}

\item{compare}{Factor to compare}

\item{ref}{Reference level}

\item{within}{Factor(s) to stratify by}

\item{method}{Analysis method}

\item{test}{Pairwise test (if applicable)}

\item{alpha}{Significance threshold}

\item{fdr_method}{FDR correction method}
}
\value{
Combined results across strata
}
\description{
Runs comparisons separately within each level of stratifying factor(s).
}
\keyword{internal}
