% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepdiff.R
\name{count_peptides_needing_m_reps}
\alias{count_peptides_needing_m_reps}
\title{Count Peptides Needing a give Number of Replicates}
\usage{
count_peptides_needing_m_reps(r, m = 10, b = 0.8)
}
\arguments{
\item{r}{results usually from `compare_many()`  containing information about peptide comparisons.}

\item{m}{The minimum number of replicates required for a peptide at the given power.}

\item{b}{The power threshold for categorizing comparisons ("Sufficient Power" or "Under Power").}
}
\value{
A data frame with counts of peptides for each comparison and power category.
}
\description{
This function takes results object and counts the number of peptides
that need fewer than a specified minimum number of replicates for the given power.
}
