% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{drop_columns}
\alias{drop_columns}
\title{Drop Columns from a Dataframe Based on Significance and Metric}
\usage{
drop_columns(df, sig, metric, log, base, rows_to_keep = NULL)
}
\arguments{
\item{df}{The input dataframe.}

\item{sig}{Significance cutoff to filter rows.}

\item{metric}{The metric used for filtering columns.}

\item{log}{Logical, indicating whether to log-transform the fold_change column.}

\item{base}{The base for logarithmic transformation.}

\item{rows_to_keep}{Optionally, specific rows to keep in the dataframe.}
}
\value{
A modified dataframe with selected columns.
}
\description{
This function takes a dataframe, removes unnecessary columns based on a significance cutoff and metric,
and returns a modified dataframe.
}
\examples{
\dontrun{
# Example Usage:
result_df <- drop_columns(my_dataframe, sig = 0.05,
metric = "bootstrap_t_p_val", log = TRUE, base = 2)
}
}
