% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{estimate_result_clusters}
\alias{estimate_result_clusters}
\title{plots a Figure of Merit curve to help estimate the number of clusters in the results}
\usage{
estimate_result_clusters(r)
}
\arguments{
\item{r}{the results object from `compare_many()`}
}
\description{
plots a Figure of Merit curve to help estimate the number of clusters in the results
}
