% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{extract_stored_residuals}
\alias{extract_stored_residuals}
\title{Extract stored residuals and fitted values for selected peptides}
\usage{
extract_stored_residuals(diagnostics, peptides)
}
\description{
Extract stored residuals and fitted values for selected peptides
}
\keyword{internal}
