% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{fc_qqplot}
\alias{fc_qqplot}
\title{plot qqplot of fold changes from a comparison}
\usage{
fc_qqplot(df, log = FALSE, base = 2)
}
\arguments{
\item{df}{result dataframe, typically from `compare()`}

\item{log}{log the fold change values}

\item{base}{base of the log, if used}
}
\description{
plot qqplot of fold changes from a comparison
}
