% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fit_and_extract_art}
\alias{fit_and_extract_art}
\title{Fit ART and extract contrasts for one peptide}
\usage{
fit_and_extract_art(data, response, factors, compare, peptide_id)
}
\arguments{
\item{data}{Data for one peptide}

\item{response}{Name of the response variable}

\item{factors}{Character vector of factor names}

\item{compare}{Which factor to compare}

\item{peptide_id}{Peptide identifier}
}
\value{
A list with model fit results and contrasts
}
\description{
Fit ART and extract contrasts for one peptide
}
\keyword{internal}
