% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{get_all_residuals}
\alias{get_all_residuals}
\title{Get standardized residuals from all peptide models}
\usage{
get_all_residuals(results)
}
\description{
Get standardized residuals from all peptide models
}
\keyword{internal}
