% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{get_comparison}
\alias{get_comparison}
\title{Get results for a specific comparison}
\usage{
get_comparison(x, comparison)
}
\arguments{
\item{x}{A pepdiff_results object}

\item{comparison}{Comparison name to retrieve}
}
\value{
A tibble with results for the specified comparison
}
\description{
Get results for a specific comparison
}
