% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_eb}
\alias{get_eb}
\title{Calculate Empirical Bayes Statistics for Treatment vs. Control Comparison}
\usage{
get_eb(treatment, control)
}
\arguments{
\item{treatment}{A matrix or data frame representing the treatment group's data.}

\item{control}{A matrix or data frame representing the control group's data.}
}
\value{
A data frame containing empirical Bayes p-values, FDR-adjusted p-values, and Vash statistics.
}
\description{
This function computes empirical Bayes statistics for comparing treatment and control groups.
}
\details{
The function fits a linear model using limma::lmFit and performs empirical Bayes moderation with limma::eBayes.
The Vashr package is used to compute additional statistics for the comparison.
}
\examples{
\dontrun{
treatment <- matrix(rnorm(100), ncol = 5)
control <- matrix(rnorm(100), ncol = 5)
results <- get_eb(treatment, control)
}

}
\references{
Ritchie, M. E., Phipson, B., Wu, D., Hu, Y., Law, C. W., Shi, W., & Smyth, G. K. (2015).
limma powers differential expression analyses for RNA-sequencing and microarray studies.
}
