% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_gamma}
\alias{get_gamma}
\title{Calculate Gamma Values for Peptide Data}
\usage{
get_gamma(treatment, control)
}
\arguments{
\item{treatment}{A matrix or data frame representing the treatment group's data.}

\item{control}{A matrix or data frame representing the control group's data.}
}
\value{
A data frame containing gamma p-values and FDR-adjusted gamma values for each peptide.
}
\description{
This function computes gamma values for each peptide based on a treatment and control group.
}
\details{
The function fits a Generalized Linear Model (GLM) using the Gamma family to compare treatment and control groups for each peptide.
}
\examples{
\dontrun{
treatment <- matrix(rnorm(100), ncol = 5)
control <- matrix(rnorm(100), ncol = 5)
results <- get_gamma(treatment, control)
}
}
\references{
Davison, A. C., & Hinkley, D. V. (1997). Bootstrap Methods and Their Application.
}
\seealso{
\code{\link{glm}}, \code{\link{p.adjust}}
}
