% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_kruskal_percentile}
\alias{get_kruskal_percentile}
\title{Calculate Kruskal-Wallis test p-values and adjusted false discovery rates for 2 peptide comparisons.}
\usage{
get_kruskal_percentile(treatment, control)
}
\arguments{
\item{treatment}{A dataframe representing the treatment group with quantitative values for each peptide.}

\item{control}{A dataframe representing the control group with quantitative values for each peptide.}
}
\value{
A dataframe containing Kruskal-Wallis test p-values and corresponding Bonferroni-adjusted FDR values for each peptide comparison.
}
\description{
This function computes Kruskal-Wallis test p-values for each peptide comparison between treatment and control groups. It also calculates adjusted false discovery rates (FDR) using the Bonferroni method. The Kruskal-Wallis test is used to assess differences in the distributions of quantitative values across multiple groups.
}
\examples{
\dontrun{
# Example using two dataframes for treatment and control:
treatment_data <- data.frame(peptide1 = c(10, 20, 30),
peptide2 = c(15, 25, 35))
control_data <- data.frame(peptide1 = c(12, 22, 32),
peptide2 = c(18, 28, 38))
result <- get_kruskal_percentile(treatment_data, control_data)
}
}
\seealso{
Other statistical analysis: 
\code{\link{compare_many}()},
\code{\link{compare}()},
\code{\link{get_bootstrap_percentile}()},
\code{\link{get_percentile_lowest_observed_value_iterative}()},
\code{\link{get_rp_percentile}()},
\code{\link{get_wilcoxon_percentile}()}
}
\concept{statistical analysis}
\keyword{data}
