% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{get_peptide}
\alias{get_peptide}
\title{Get results for a specific peptide}
\usage{
get_peptide(x, peptide)
}
\arguments{
\item{x}{A pepdiff_results object}

\item{peptide}{Peptide ID to retrieve}
}
\value{
A tibble with results for the specified peptide
}
\description{
Get results for a specific peptide
}
