% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepdiff.R
\name{get_percentile_iterative}
\alias{get_percentile_iterative}
\title{first developed method for getting the p estimate}
\usage{
get_percentile_iterative(d, global_mean, global_sd, iters = 1000)
}
\description{
first developed method for getting the p estimate
}
