% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_power}
\alias{get_power}
\title{do a power analysis for each peptide}
\usage{
get_power(treatment, control, sig_level = 0.05, b = 0.8, max_n = 50)
}
\arguments{
\item{treatment}{matrix of treatment data columns}

\item{control}{matrix of control data columns}

\item{sig_level}{p-value expected to be used in statistical tests}

\item{b}{minimal power of test expected}

\item{max_n}{maximum number of reps to try when searching for replicates needed to reach power of b}
}
\description{
takes the treatment and control values for a peptide and
assesses the probability of finding a difference if one exists at the
calculated Cohen's d effect size, the standard deviation and sample size for
these data. The power required for a two-tailed t-test, Cohen's d and an
an estimate of the number of replicates required to reach a given proabability
of detecting a difference at that effect size are returned
}
