% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_sig_rows}
\alias{get_sig_rows}
\title{works out if a row has at least one significant value in it}
\usage{
get_sig_rows(l, metric = "bootstrap_t_pval", sig_level = 0.05)
}
\description{
works out if a row has at least one significant value in it
}
