% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{has_sufficient_variation}
\alias{has_sufficient_variation}
\title{Check if a model has enough variation to fit}
\usage{
has_sufficient_variation(values)
}
\arguments{
\item{values}{Numeric vector of values}
}
\value{
Logical
}
\description{
Check if a model has enough variation to fit
}
\keyword{internal}
