% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepdiff.R
\name{is_useable}
\alias{is_useable}
\title{calculate the number of non- NA or NaNs in a vector}
\usage{
is_useable(x)
}
\arguments{
\item{x}{vector of values}
}
\value{
number
}
\description{
calculate the number of non- NA or NaNs in a vector
}
