% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{jzs_bf_approx}
\alias{jzs_bf_approx}
\title{JZS Bayes factor approximation}
\usage{
jzs_bf_approx(t_stat, n_eff, r_scale = 0.707)
}
\arguments{
\item{t_stat}{T-statistic}

\item{n_eff}{Effective sample size}

\item{r_scale}{Scale parameter for Cauchy prior}
}
\value{
Bayes factor (BF10)
}
\description{
Computes BF10 using the Savage-Dickey density ratio approximation.
}
\keyword{internal}
