% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{kmeans_by_selected_cols}
\alias{kmeans_by_selected_cols}
\title{Perform kmeans of a dataset using just data in selected columns, then return matrices of all columns}
\usage{
kmeans_by_selected_cols(
  l,
  cols = NULL,
  log = TRUE,
  base = 2,
  sig_only = TRUE,
  sig_level = 0.05,
  metric = "bootstrap_t_p_val",
  k = NA,
  nstart = 25,
  itermax = 1000
)
}
\arguments{
\item{l}{list of results, usually from `compare_many()`}

\item{cols}{names of columns to perform the k-means with}

\item{log}{whether to log the data}

\item{base}{base used in logging (default = 2)}

\item{sig_only}{return only rows with 1 or more values significant at `sig_level` of `metric`}

\item{sig_level}{significance level cutoff}

\item{metric}{the test metric used to determine significance one of:
`bootstrap_t_p_val`, `bootstrap_t_fdr`
`wilcoxon_p_val`, `wilcoxon_fdr`
`kruskal_p_val`,  `kruskal_fdr`
`rank_prod_p1_p_val`, `rank_prod_p2_p_val`, `rank_prod_p1_fdr`, `rank_prod_p2_fdr`.}

\item{k}{number of clusters to make}

\item{nstart}{nstart value for `kmeans()`}

\item{itermax}{number of `kmeans()` iterations (1000)}
}
\value{
list of matrices
}
\description{
Perform kmeans of a dataset using just data in selected columns, then return matrices of all columns
}
