% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log2_fc}
\alias{log2_fc}
\title{Calculate log2 fold change}
\usage{
log2_fc(fold_change)
}
\arguments{
\item{fold_change}{Numeric vector of fold changes}
}
\value{
Numeric vector of log2 fold changes
}
\description{
Calculate log2 fold change
}
\keyword{internal}
