% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{long_results}
\alias{long_results}
\title{convert wide format results table containing p-value estimates to long format}
\usage{
long_results(r)
}
\arguments{
\item{r}{results dataframe typically from `compare()`}
}
\value{
dataframe in long format missing some columns from r
}
\description{
tidies up the wide results table from `compare()` to a long format, dropping the fdr columns and quantitie columns
}
