% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mean_fold_change}
\alias{mean_fold_change}
\title{calculate mean fold change for peptide}
\usage{
mean_fold_change(treatment, control)
}
\arguments{
\item{treatment}{vector of `treatment` data}

\item{control}{vector of `control` data}
}
\value{
vector of mean fold changes `mean(treatment) / mean(control)`
}
\description{
for each peptide calculate the mean quantification over all experiments
then get the natural scale fold change.
}
