% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{min_peptide_values}
\alias{min_peptide_values}
\title{find minimal values for a peptide across replicates}
\usage{
min_peptide_values(d)
}
\arguments{
\item{d}{matrix of peptide quantifications. rows are peptides, columns are
experimental conditions}
}
\value{
vector of minimume values of length nrow(d)
}
\description{
find minimal values for a peptide across replicates
}
