% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepdiff.R
\name{missing_peptides_plot}
\alias{missing_peptides_plot}
\title{plot the representation of peptides in each group.}
\usage{
missing_peptides_plot(df)
}
\arguments{
\item{df}{dataframe with unmerged tech reps; typically from `import_data()`}
}
\value{
ggplot2 plot
}
\description{
Shows what proportion of the whole set of peptides is missing in each group
of treatment, seconds, bio rep and tech rep.
}
