% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{p_value_hist}
\alias{p_value_hist}
\title{plot histograms of p-values for each test used}
\usage{
p_value_hist(r)
}
\arguments{
\item{df}{result dataframe typically from `compare()`}
}
\value{
ggplot2 plot
}
\description{
plot histograms of p-values for each test used
}
