% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.pepdiff_results}
\alias{plot.pepdiff_results}
\title{Plot method for pepdiff_results}
\usage{
\method{plot}{pepdiff_results}(x, ...)
}
\arguments{
\item{x}{A pepdiff_results object}

\item{...}{Additional arguments (ignored)}
}
\value{
A cowplot grid of plots
}
\description{
Creates a multi-panel plot showing volcano, p-value histogram, and FC distribution.
}
