% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_cluster_estimate}
\alias{plot_cluster_estimate}
\title{plots a Figure of Merit curve to help estimate the number of clusters in the results for a given metric and a
given significance level. Clustering is done based on Fold}
\usage{
plot_cluster_estimate(
  r,
  log = TRUE,
  base = 2,
  sig_only = FALSE,
  sig_level = 0.05,
  metric = "bootstrap_t_fdr"
)
}
\arguments{
\item{r}{the results object from `compare()` or `compare_many()`}

\item{log}{log the fold changes}

\item{base}{base for logging}

\item{sig_level}{significance cutoff for colour}

\item{metric}{metric to use for significance}
}
\description{
plots a Figure of Merit curve to help estimate the number of clusters in the results for a given metric and a
given significance level. Clustering is done based on Fold
}
