% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_compared_calls}
\alias{plot_compared_calls}
\title{Compare Sets of Significant Peptides Called by the Used Data}
\usage{
plot_compared_calls(r, sig = 0.05)
}
\arguments{
\item{r}{A result dataframe or list typically obtained from `compare()` or `compare_many()`.}

\item{sig}{Significance cutoff to select peptides.}
}
\value{
UpSet plot.
}
\description{
Produces an UpSet plot showing intersections and set-size of the different
sets of significant peptides called by the methods used in the provided result
dataframe.
}
\examples{
\dontrun{
plot_compared_calls(my_results, sig = 0.05)
}
}
