% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_distributions_simple}
\alias{plot_distributions_simple}
\title{Simple distribution plot for pepdiff_data}
\usage{
plot_distributions_simple(data, facet_by = NULL)
}
\arguments{
\item{data}{A pepdiff_data object}

\item{facet_by}{Factor to facet by (default: first factor)}
}
\value{
A ggplot object
}
\description{
Simple distribution plot for pepdiff_data
}
