% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_effect_replicates}
\alias{plot_effect_replicates}
\title{Plot Effect of Replicates}
\usage{
plot_effect_replicates(r, which = "cohens_d", b = 0.8)
}
\arguments{
\item{r}{Dataframe or list of result dataframes typically from `compare()` or `compare_many()`.}

\item{which}{Indicates whether to use "cohens_d" or "fold_change" for the x-axis.}

\item{b}{Threshold for power to categorize as "Sufficient Power" or "Under Power".}
}
\description{
This function plots the effect of replicates based on either Cohen's D or Fold Change.
}
\examples{
\dontrun{
# Example Usage:
plot_effect_replicates(my_results, which = "cohens_d", b = 0.8)
}
}
