% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_fc_distribution_new}
\alias{plot_fc_distribution_new}
\title{Fold change distribution for pepdiff_results}
\usage{
plot_fc_distribution_new(results, comparison = NULL)
}
\arguments{
\item{results}{A pepdiff_results object}

\item{comparison}{Optional comparison to filter by}
}
\value{
A ggplot object
}
\description{
Fold change distribution for pepdiff_results
}
