% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_fdr_dist}
\alias{plot_fdr_dist}
\title{Plot Distribution of FDR Values}
\usage{
plot_fdr_dist(r)
}
\arguments{
\item{r}{A data frame containing FDR values for different comparisons and tests.}
}
\value{
A ggplot object displaying the distribution of FDR values.
}
\description{
This function generates a histogram to visualize the distribution of False Discovery Rate (FDR) values
for different comparisons and tests.
}
\details{
The function uses ggplot2 to create a histogram, with facets for different comparisons and test
}
