% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_fdr_qq}
\alias{plot_fdr_qq}
\title{Plot Histograms of False Discovery Rates (FDR)}
\usage{
plot_fdr_qq(r)
}
\arguments{
\item{r}{A dataframe or list of result dataframes.}
}
\value{
A ggplot2 plot showing histograms of False Discovery Rates (FDR) for each test.
}
\description{
This function takes a dataframe or a list of result dataframes, typically obtained from
functions like `compare()` or `compare_many()`, and plots histograms of False Discovery Rates (FDR) for each test.
}
\examples{
\dontrun{
plot_fdr_dist(my_results)
}
}
