% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_kmeans}
\alias{plot_kmeans}
\title{K-means cluster the data on the samples}
\usage{
plot_kmeans(df, nstart = 25, iter.max = 1000)
}
\arguments{
\item{df}{dataframe, typically from `import_data()`}

\item{nstart}{nstart points for `kmeans()` function}

\item{iter.max}{max iterations to perform for `kmeans()` function}
}
\value{
ggplot2 plot
}
\description{
Performs and draws a K-means cluster on the samples. Estimates number of clusters
as the product of the number of treatments and seconds. So tries to group the bio reps together
}
