% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_kmeans_cluster_hmap}
\alias{plot_kmeans_cluster_hmap}
\title{plots the kmeans clusters}
\usage{
plot_kmeans_cluster_hmap(
  kl,
  col_order = NULL,
  logged = TRUE,
  base = NULL,
  pal = "RdBu",
  nrow = 2,
  col_fontsize = 6,
  row_fontsize = 6,
  legx = 0,
  legy = -0.45,
  labh = -1.1,
  labv = 1.1
)
}
\arguments{
\item{kl}{the kmeans clusters from `kmeans_by_selected_cols()`}

\item{col_order}{vector of column names in the order you want them to be drawn}

\item{logged}{indicates whether the fold changes are logged or not}

\item{base}{if logged the base that was used}

\item{pal}{pallete to use from RColorBrewer}

\item{nrow}{number of rows to plot clusters in}

\item{legx}{x position of legend in range (-0.5 .. 0.5)}

\item{legy}{y position of legend in range(-0.5 .. 0.5)}

\item{labh}{horizontal justification of subplot labels}

\item{labv}{vertical justification of subplot labels}
}
\description{
plots the kmeans clusters
}
