% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_missingness_simple}
\alias{plot_missingness_simple}
\title{Simple missingness plot for pepdiff_data}
\usage{
plot_missingness_simple(data)
}
\arguments{
\item{data}{A pepdiff_data object}
}
\value{
A ggplot object
}
\description{
Simple missingness plot for pepdiff_data
}
