% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_norm_qq}
\alias{plot_norm_qq}
\title{draw qqplots for data}
\usage{
plot_norm_qq(i, log = FALSE, base = 2)
}
\arguments{
\item{i}{dataframe; typically from `import_data()`}

\item{log}{perform log transform of data}

\item{base}{base to use in log transform}
}
\value{
ggplot2 plot
}
\description{
Plot qqplot of distribution of quantifications in data for each treatment,
seconds and biological replicate
}
