% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_p_value_dist}
\alias{plot_p_value_dist}
\title{plot histograms of p-values for each test used}
\usage{
plot_p_value_dist(r)
}
\arguments{
\item{r}{dataframe or list of result dataframes typically from `compare()` or `compare_many()`}
}
\value{
ggplot2 plot
}
\description{
plot histograms of p-values for each test used
}
