% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_p_value_qq}
\alias{plot_p_value_qq}
\title{Plot Quantile-Quantile (QQ) Plot of p-values}
\usage{
plot_p_value_qq(r)
}
\arguments{
\item{r}{A dataframe or list of result dataframes.}
}
\value{
A ggplot2 QQ plot showing the distribution of p-values for each test.
}
\description{
This function takes a dataframe or a list of result dataframes, typically obtained from
functions like `compare()` or `compare_many()`, and generates QQ plots of p-values for each test.
}
\examples{
\dontrun{
plot_p_value_qq(my_results)
}

}
