% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_pca}
\alias{plot_pca}
\title{Create a PCA (Principal Component Analysis) plot and screeplot for multivariate data.}
\usage{
plot_pca(df)
}
\arguments{
\item{df}{A data frame containing multivariate data.}
}
\value{
A combined ggplot object representing the PCA plot and screeplot.
}
\description{
This function performs Principal Component Analysis (PCA) on input data and generates a PCA plot and a screeplot. The PCA plot visualizes the data points in a two-dimensional space defined by the first two principal components, and it is  colored by biological replicate, treatment, or time point in each subplot. The screeplot shows the proportion of variance explained by each principal component.
}
\examples{
\dontrun{
plot_pca(data)
}
}
\seealso{
Other data visualization: 
\code{\link{plot_missing_peptides}()},
\code{\link{plot_norm_qq}()},
\code{\link{plot_peptides_measured}()},
\code{\link{plot_quant_distributions}()},
\code{\link{plot_sample_kmeans}()}
}
\concept{data visualization}
\keyword{plot}
