% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_pca_simple}
\alias{plot_pca_simple}
\title{Simple PCA plot for pepdiff_data}
\usage{
plot_pca_simple(data, color_by = NULL)
}
\arguments{
\item{data}{A pepdiff_data object}

\item{color_by}{Factor to color points by (default: first factor)}
}
\value{
A ggplot object
}
\description{
Simple PCA plot for pepdiff_data
}
