% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_power_analysis}
\alias{plot_power_analysis}
\title{Power Analysis Plot}
\usage{
plot_power_analysis(r)
}
\arguments{
\item{r}{Dataframe or list of result dataframes typically from `compare()` or `compare_many()`.}
}
\description{
This function plots a power analysis histogram for a given comparison.
}
\examples{
\dontrun{
# Example Usage:
plot_power_analysis(my_results)
}
}
