% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_power_volcano}
\alias{plot_power_volcano}
\title{Power Volcano Plot}
\usage{
plot_power_volcano(
  r,
  which = "fold_change",
  base = 2,
  b = 0.8,
  option = "E",
  direction = -1
)
}
\arguments{
\item{r}{Dataframe or list of result dataframes typically from `compare()` or `compare_many()`.}

\item{which}{Indicates whether to use "fold_change" or "cohens_d" for the x-axis.}

\item{base}{The base for logarithmic transformation if using "fold_change".}

\item{b}{Threshold for power to categorize as "Sufficient Power" or "Under Power".}

\item{option}{Color scale option for the viridis palette.}

\item{direction}{Direction of the viridis color scale.}
}
\description{
This function plots a volcano plot showing the relationship between fold change, power, and significance.
}
\examples{
\dontrun{
# Example Usage:
plot_power_volcano(my_results, which = "fold_change", base = 2,
b = 0.8, option = "E", direction = -1)
}
}
