% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_pvalue_histogram}
\alias{plot_pvalue_histogram}
\title{P-value histogram for pepdiff_results}
\usage{
plot_pvalue_histogram(results, comparison = NULL)
}
\arguments{
\item{results}{A pepdiff_results object}

\item{comparison}{Optional comparison to filter by}
}
\value{
A ggplot object
}
\description{
P-value histogram for pepdiff_results
}
