% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_result}
\alias{plot_result}
\title{plot the p-values against fold change for the tests used in `compare()`}
\usage{
plot_result(df)
}
\arguments{
\item{df}{result dataframe typically from `compare()`}
}
\value{
ggplot2 plot
}
\description{
plots fold change against p-value in all tests used, also splits data on the number
of biological replicates were available before value replacement for each peptide
}
