% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_sample_kmeans}
\alias{plot_sample_kmeans}
\title{Apply k-Means clustering to sample data and visualize the clustering results.}
\usage{
plot_sample_kmeans(df, nstart = 25, iter.max = 1000)
}
\arguments{
\item{df}{A data frame containing multivariate data.}

\item{nstart}{The number of random initializations for the k-Means algorithm. Default is 25.}

\item{iter.max}{The maximum number of iterations for the k-Means algorithm. Default is 1000.}
}
\value{
A ggplot object representing the k-Means clustering results.
}
\description{
This function applies k-Means clustering to the first three principal components obtained from PCA analysis of the data. k is estimated from the sample It then visualizes the clustering results by creating a scatterplot of the PCA components, color-coded according to the k-Means clusters the samples fall into.
}
\examples{
\dontrun{
plot_sample_kmeans(data, nstart = 25, iter.max = 1000)
}
}
\seealso{
Other data visualization: 
\code{\link{plot_missing_peptides}()},
\code{\link{plot_norm_qq}()},
\code{\link{plot_pca}()},
\code{\link{plot_peptides_measured}()},
\code{\link{plot_quant_distributions}()}
}
\concept{data visualization}
\keyword{plot}
