% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{print.pepdiff_data}
\alias{print.pepdiff_data}
\title{Print method for pepdiff_data}
\usage{
\method{print}{pepdiff_data}(x, ...)
}
\arguments{
\item{x}{A pepdiff_data object}

\item{...}{Additional arguments (ignored)}
}
\value{
The object invisibly
}
\description{
Print method for pepdiff_data
}
