% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{print.pepdiff_results}
\alias{print.pepdiff_results}
\title{Print method for pepdiff_results}
\usage{
\method{print}{pepdiff_results}(x, ...)
}
\arguments{
\item{x}{A pepdiff_results object}

\item{...}{Additional arguments (ignored)}
}
\value{
The object invisibly
}
\description{
Print method for pepdiff_results
}
