% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{print_diagnostic_summary}
\alias{print_diagnostic_summary}
\title{Print diagnostic summary to console}
\usage{
print_diagnostic_summary(summary_stats)
}
\description{
Print diagnostic summary to console
}
\keyword{internal}
