% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{refit_and_extract_residuals}
\alias{refit_and_extract_residuals}
\title{Refit GLM models and extract residuals for selected peptides}
\usage{
refit_and_extract_residuals(results, peptides)
}
\description{
Refit GLM models and extract residuals for selected peptides
}
\keyword{internal}
