% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{replace_vals}
\alias{replace_vals}
\title{Replace missing values with replacements (legacy)}
\usage{
replace_vals(x, lowest_vals)
}
\arguments{
\item{x}{Vector with missing values}

\item{lowest_vals}{Vector of replacements}
}
\value{
Vector with values replaced
}
\description{
Replace missing values with replacements (legacy)
}
\keyword{internal}
