% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{run_models}
\alias{run_models}
\title{Run statistical model for all peptides}
\usage{
run_models(data, compare, ref = NULL, method = "glm")
}
\arguments{
\item{data}{A pepdiff_data object}

\item{compare}{Factor to compare}

\item{ref}{Reference level}

\item{method}{One of "glm" or "art"}
}
\value{
A list with results for each peptide
}
\description{
Applies GLM or ART model to each peptide in the dataset.
}
\keyword{internal}
