% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{select_columns_for_contrast}
\alias{select_columns_for_contrast}
\title{Select columns for contrast (legacy)}
\usage{
select_columns_for_contrast(
  l,
  treatment = NA,
  t_seconds = NA,
  control = NA,
  c_seconds = NA
)
}
\arguments{
\item{l}{List of data and row_info}

\item{treatment}{Treatment name}

\item{t_seconds}{Treatment seconds}

\item{control}{Control name}

\item{c_seconds}{Control seconds}
}
\value{
List with treatment and control matrices
}
\description{
Select columns for contrast (legacy)
}
\keyword{internal}
