% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{subset.pepdiff_data}
\alias{subset.pepdiff_data}
\title{Subset a pepdiff_data object}
\usage{
\method{subset}{pepdiff_data}(x, peptides = NULL, ...)
}
\arguments{
\item{x}{A pepdiff_data object}

\item{peptides}{Character vector of peptide IDs to keep (optional)}

\item{...}{Additional subsetting expressions evaluated in the data}
}
\value{
A new pepdiff_data object
}
\description{
Subset a pepdiff_data object
}
