% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{summary.pepdiff_data}
\alias{summary.pepdiff_data}
\title{Summary method for pepdiff_data}
\usage{
\method{summary}{pepdiff_data}(object, ...)
}
\arguments{
\item{object}{A pepdiff_data object}

\item{...}{Additional arguments (ignored)}
}
\value{
A summary list invisibly
}
\description{
Summary method for pepdiff_data
}
