% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{summary.pepdiff_results}
\alias{summary.pepdiff_results}
\title{Summary method for pepdiff_results}
\usage{
\method{summary}{pepdiff_results}(object, ...)
}
\arguments{
\item{object}{A pepdiff_results object}

\item{...}{Additional arguments (ignored)}
}
\value{
A summary list invisibly
}
\description{
Summary method for pepdiff_results
}
