% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy-pepdiff.R
\name{times_measured}
\alias{times_measured}
\title{calculate number of measurements of each peptide in each treatment and time}
\usage{
times_measured(df)
}
\arguments{
\item{df}{dataframe. Typically from `import_data()`}
}
\value{
dataframe
}
\description{
For each peptide, works out how many biologically replicated measurements are
available in the different combinations of treatment and seconds
}
