% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{times_measured_plot}
\alias{times_measured_plot}
\title{plot the count of the number of times peptides were measured.}
\usage{
times_measured_plot(df)
}
\arguments{
\item{df}{dataframe. Typically from `import_data()`}
}
\value{
ggplot2 plot
}
\description{
Calculates and plots the number of times each peptide was measured in each
combination of treatment and seconds and presents a summary plot
}
