% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_vec}
\alias{to_vec}
\title{extract data from a loaded peptide data object}
\usage{
to_vec(i)
}
\arguments{
\item{i}{input dataframe}
}
\description{
extract peptide quantification data from a loaded data file, usually from
`import_data()` for use in `fitdistrplus` functions
}
