% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_factors}
\alias{validate_factors}
\title{Validate factor columns exist in data}
\usage{
validate_factors(data, factors)
}
\arguments{
\item{data}{A data frame}

\item{factors}{Character vector of factor column names}
}
\value{
TRUE invisibly if valid, otherwise throws an error
}
\description{
Checks that all specified factor columns exist in the data frame
and are not empty.
}
\keyword{internal}
