% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{validate_pepdiff_results}
\alias{validate_pepdiff_results}
\title{Validate a pepdiff_results object}
\usage{
validate_pepdiff_results(x)
}
\arguments{
\item{x}{A pepdiff_results object to validate}
}
\value{
The validated object (invisibly), or throws an error
}
\description{
Validate a pepdiff_results object
}
\keyword{internal}
