% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_positive}
\alias{validate_positive}
\title{Validate positive values for Gamma GLM}
\usage{
validate_positive(values, name = "values")
}
\arguments{
\item{values}{Numeric vector of values to check}

\item{name}{Name of the variable for error messages}
}
\value{
TRUE invisibly if valid, otherwise throws an error
}
\description{
Checks that all non-NA values are strictly positive (> 0).
Gamma GLM requires positive values.
}
\keyword{internal}
