% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{volcano_plot}
\alias{volcano_plot}
\title{volcano plot the data}
\usage{
volcano_plot(
  l,
  log = FALSE,
  base = 2,
  sig_level = 0.05,
  metric = "bootstrap_t_p_val",
  option = "E",
  direction = -1
)
}
\arguments{
\item{l}{list of results data frames, typically from `compare_many()`}

\item{log}{log the data}

\item{base}{base for logging}

\item{sig_level}{significance cutoff for colour}

\item{metric}{metric to use for significance}

\item{option}{viridis colour scheme key to use}

\item{direction}{viridis colour scheme direction (1/-1)}
}
\value{
ggplot2 plot
}
\description{
draws a plot of peptide count against log fc at either protein or peptide level for samples
}
