% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepdiff.R
\name{compare}
\alias{compare}
\title{compare the peptide quantities in two experiments}
\usage{
compare(
  df,
  iters = 1000,
  treatment = NA,
  t_seconds = NA,
  control = NA,
  c_seconds = NA,
  metrics = NA
)
}
\arguments{
\item{df}{dataframe. Typically from `import_data()`}

\item{iters}{number of iterations to perform for iterative tests}

\item{treatment}{name of the experimental treatment to use as the `treatment` condition}

\item{t_seconds}{time point of the `treatment` condition to use}

\item{control}{name of the experimental treatment to use as the `control` condition}

\item{c_seconds}{time point of the `control` condition to use}

\item{metrics}{character vector of tests to use, one or more of: `norm_quantile`, `bootstrap_t`, `wilcoxon`, `kruskal-wallis`, `rank_product`}
}
\value{
dataframe with original and replaced quantification values, natural fold change, biological replicates and p-value / fdr for each
}
\description{
For each peptide this function carries out the selected tests to determine peptides
that are differentially abundant in the two experiments specified. Before tests are
performed the technical replicates are merged and lowest observed value replacement
for each missing biological replicate is done. All comparisins are performed as `treatment / control`
}
