% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy-pepdiff.R
\name{compare_many}
\alias{compare_many}
\title{compare many combinations of treatment and control}
\usage{
compare_many(df, comparison, iters = 1000, tests = c("bootstrap_t"))
}
\arguments{
\item{df}{dataframe. Typically from `import_data()`}

\item{comparison}{path to file or dataframe of comparisons with columns treatment, t_seconds, control, c_seconds}

\item{iters}{number of iterations to perform for iterative tests}

\item{tests}{character vector of tests to use, one or more of: `norm_quantile`, `bootstrap_t`, `wilcoxon`, `kruskal-wallis`, `rank_product`}
}
\description{
for each combination of treatment and control condition, runs the `compare()` function
and collates the results
}
