% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_bootstrap_percentile}
\alias{get_bootstrap_percentile}
\title{get p values for contrast using boostrap t test}
\usage{
get_bootstrap_percentile(treatment, control, iters)
}
\arguments{
\item{treatment}{matrix of treatment data columns}

\item{control}{matrix of control data columns}

\item{iters}{number of iterations to perform}
}
\value{
dataframe with two columns `bootstrap_t_p_val` and `bootstrap_t_fdr`
}
\description{
get p values for contrast using boostrap t test
}
