% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy-stats.R, R/tests.R
\name{get_bootstrap_percentile}
\alias{get_bootstrap_percentile}
\title{get p values for contrast using boostrap t test}
\usage{
get_bootstrap_percentile(treatment, control, iters = 1000)

get_bootstrap_percentile(treatment, control, iters = 1000)
}
\arguments{
\item{treatment}{Matrix of treatment data (rows = peptides, cols = replicates)}

\item{control}{Matrix of control data}

\item{iters}{Number of bootstrap iterations}
}
\value{
dataframe with two columns `bootstrap_t_p_val` and `bootstrap_t_fdr`

Data frame with bootstrap_t_p_val and bootstrap_t_fdr columns
}
\description{
get p values for contrast using boostrap t test

Legacy: Get bootstrap t-test p-values for matrix data
}
