% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy-stats.R, R/tests.R
\name{get_kruskal_percentile}
\alias{get_kruskal_percentile}
\title{get p values for contrast using Kruskal-Wallis test}
\usage{
get_kruskal_percentile(treatment, control)

get_kruskal_percentile(treatment, control)
}
\arguments{
\item{treatment}{Matrix of treatment data}

\item{control}{Matrix of control data}
}
\value{
dataframe with two columns `kruskal_p_val` and `kruskal_fdr`

Data frame with kruskal_p_val and kruskal_fdr columns
}
\description{
get p values for contrast using Kruskal-Wallis test

Legacy: Get Kruskal-Wallis test p-values for matrix data
}
