% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_kruskal_percentile}
\alias{get_kruskal_percentile}
\title{get p values for contrast using Kruskal-Wallis test}
\usage{
get_kruskal_percentile(treatment, control)
}
\arguments{
\item{treatment}{matrix of treatment data columns}

\item{control}{matrix of control data columns}
}
\value{
dataframe with two columns `kruskal_p_val` and `kruskal_fdr`
}
\description{
get p values for contrast using Kruskal-Wallis test
}
