% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_percentile_lowest_observed_value_iterative}
\alias{get_percentile_lowest_observed_value_iterative}
\title{get p values for contrast using normal percentile}
\usage{
get_percentile_lowest_observed_value_iterative(
  treatment,
  control,
  iters = 1000
)
}
\arguments{
\item{treatment}{matrix of treatment data columns}

\item{control}{matrix of control data columns}

\item{iters}{number of iterations to perform}
}
\value{
dataframe with one column `norm_quantile_pval`
}
\description{
get p values for contrast using normal percentile
}
