% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy-stats.R, R/legacy.R
\name{get_percentile_lowest_observed_value_iterative}
\alias{get_percentile_lowest_observed_value_iterative}
\title{get p values for contrast using normal percentile}
\usage{
get_percentile_lowest_observed_value_iterative(
  treatment,
  control,
  iters = 1000
)

get_percentile_lowest_observed_value_iterative(
  treatment,
  control,
  iters = 1000
)
}
\arguments{
\item{treatment}{Treatment matrix}

\item{control}{Control matrix}

\item{iters}{Number of iterations}
}
\value{
dataframe with one column `norm_quantile_pval`

Data frame with p-values
}
\description{
get p values for contrast using normal percentile

Get p-values using normal percentile (legacy)
}
\keyword{internal}
