% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy-stats.R, R/tests.R
\name{get_rp_percentile}
\alias{get_rp_percentile}
\title{get p values for contrast using Rank Products test}
\usage{
get_rp_percentile(treatment, control)

get_rp_percentile(treatment, control)
}
\arguments{
\item{treatment}{Matrix of treatment data}

\item{control}{Matrix of control data}
}
\value{
dataframe with four columns, two for the test each way from RankProducts `rank_prod_p1_p_val`, `rank_prod_p2_p_val` and `rank_prod_p1_fdr`, `rank_prod_p2_fdr`.

Data frame with rank product p-values and FDR
}
\description{
get p values for contrast using Rank Products test

Legacy: Get Rank Products test p-values for matrix data
}
