% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_wilcoxon_percentile}
\alias{get_wilcoxon_percentile}
\title{get p values for contrast using Wilcoxon test}
\usage{
get_wilcoxon_percentile(treatment, control)
}
\arguments{
\item{treatment}{matrix of treatment data columns}

\item{control}{matrix of control data columns}
}
\value{
dataframe with two columns `wilcoxon_p_val` and `wilcoxon_fdr`
}
\description{
get p values for contrast using Wilcoxon test
}
