% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy-stats.R, R/tests.R
\name{get_wilcoxon_percentile}
\alias{get_wilcoxon_percentile}
\title{get p values for contrast using Wilcoxon test}
\usage{
get_wilcoxon_percentile(treatment, control)

get_wilcoxon_percentile(treatment, control)
}
\arguments{
\item{treatment}{Matrix of treatment data}

\item{control}{Matrix of control data}
}
\value{
dataframe with two columns `wilcoxon_p_val` and `wilcoxon_fdr`

Data frame with wilcoxon_p_val and wilcoxon_fdr columns
}
\description{
get p values for contrast using Wilcoxon test

Legacy: Get Wilcoxon test p-values for matrix data
}
