% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepdiff.R
\name{health}
\alias{health}
\title{summarize the health of the experiment as a whole}
\usage{
health(r, b = 0.8)
}
\arguments{
\item{r}{the results object, from `compare()` or `compare_many()`}

\item{b}{the statistical power at which to evaluate}
}
\value{
i
}
\description{
returns statistics on the probability of detecting effects on at least half
the outcomes and completeness of sampling
over threshold b
}
