% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepdiff.R
\name{import_data}
\alias{import_data}
\title{read data from a file}
\usage{
import_data(
  file,
  treatment = "genotype",
  bio_rep = "bio_rep",
  tech_rep = "tech_rep",
  quant = "total_area",
  seconds = "seconds",
  gene_id = "gene_id",
  peptide = "peptide_sequence"
)
}
\arguments{
\item{file}{Path to the file to load - must be a csv file}

\item{treatment}{Column containing the treatment of the observation}

\item{bio_rep}{Column containing the biological replicate of the observation}

\item{tech_rep}{Column containing the technical replicate of the observation}

\item{quant}{Column containing the quantitation data}

\item{seconds}{Column containing timepoint of observation}

\item{gene_id}{Column containing the id of the gene this hit}

\item{peptide_sequence}{Column containing the sequence of this peptide}
}
\value{
tibble with columns gene_id, peptide, treatment, bio_rep, tech_rep, quant
}
\description{
reads data, renames columns appropriately, discards unused columns, factors and
reorders, discards duplicate rows
}
