% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_useable}
\alias{is_useable}
\title{Check if values are useable (non-NA and non-NaN)}
\usage{
is_useable(x)
}
\arguments{
\item{x}{Vector of values}
}
\value{
Logical vector indicating positions of useable values
}
\description{
Check if values are useable (non-NA and non-NaN)
}
\keyword{internal}
