% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{list2mat}
\alias{list2mat}
\title{converts a results object to a matrix as if for direct use in external heatmap functions}
\usage{
list2mat(r)
}
\description{
converts a results object to a matrix as if for direct use in external heatmap functions
}
