% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{long_results}
\alias{long_results}
\title{Convert wide format results table to long format}
\usage{
long_results(r)
}
\arguments{
\item{r}{Results dataframe typically from `compare()`}
}
\value{
Dataframe in long format
}
\description{
Tidies up the wide results table from `compare()` to a long format.
}
