% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matrix_data}
\alias{matrix_data}
\title{convert dataframe to matrix}
\usage{
matrix_data(df)
}
\arguments{
\item{df}{dataframe, typically from `import_data()`}
}
\value{
list with members `row_info` - gene ID and peptide sequence and `data`
a matrix version of the data in df
}
\description{
convert dataframe to matrix
}
