% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mean_fold_change}
\alias{mean_fold_change}
\title{Calculate mean fold change for peptide}
\usage{
mean_fold_change(treatment, control)
}
\arguments{
\item{treatment}{Vector or matrix of treatment data}

\item{control}{Vector or matrix of control data}
}
\value{
Vector of mean fold changes: mean(treatment) / mean(control)
}
\description{
For each peptide calculate the mean quantification over all experiments
then get the natural scale fold change.
}
\keyword{internal}
