% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{min_peptide_values}
\alias{min_peptide_values}
\title{Find minimal values for a peptide across replicates (legacy)}
\usage{
min_peptide_values(d)
}
\arguments{
\item{d}{Matrix of peptide quantifications}
}
\value{
Vector of minimum values
}
\description{
Find minimal values for a peptide across replicates (legacy)
}
\keyword{internal}
