% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_cluster_estimate}
\alias{plot_cluster_estimate}
\title{Plot Cluster Estimate for Figure of Merit}
\usage{
plot_cluster_estimate(
  r,
  log = TRUE,
  base = 2,
  sig_only = FALSE,
  sig_level = 0.05,
  metric = "bootstrap_t_fdr"
)
}
\arguments{
\item{r}{The results object from `compare()` or `compare_many()`.}

\item{log}{Logical, indicating whether to log-transform the fold changes.}

\item{base}{The base for logarithmic transformation.}

\item{sig_only}{Logical, indicating whether to consider only significant values.}

\item{sig_level}{Significance cutoff for color.}

\item{metric}{The metric to use for significance.}
}
\description{
This function plots a Figure of Merit curve to help estimate the number of clusters in the results for a given metric and significance level.
Clustering is done based on Fold Change.
}
\examples{
\dontrun{
# Example Usage:
plot_cluster_estimate(my_results, log = TRUE, base = 2, sig_level = 0.05,
metric = "bootstrap_t_fdr")
}
}
