% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_fc}
\alias{plot_fc}
\title{plot histogram of fold change distribution for a comparison}
\usage{
plot_fc(df, log = FALSE, base = 2)
}
\arguments{
\item{df}{result dataframe, typically from `compare()`}

\item{log}{log the fold change values}

\item{base}{base of the log, if used}
}
\value{
ggplot2 plot
}
\description{
plot histogram of fold change distribution for a comparison
}
