% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{makes heatmap from all experiments, pass a single metric and sig value
reduces dataframes and makeslong list, makes a heatmap}
\usage{
plot_heatmap(
  l,
  sig = 0.05,
  metric = NA,
  log = FALSE,
  base = 2,
  col_order = NULL,
  rotate_x_labels = TRUE
)
}
\arguments{
\item{l}{list of results data frames, typically from `compare_many()`}

\item{sig}{significance level at which to call peptides}

\item{metric}{one metric to use for the source of p-values}

\item{log}{log the fold change values to show in the heatmap}

\item{base}{base to use for logs}

\item{col_order}{specify a column order for the plot, default is names(l)}

\item{rotate_x_labels}{make the X-axis labels stand on end}
}
\value{
ggplot2 plot
}
\description{
makes heatmap from all experiments, pass a single metric and sig value
reduces dataframes and makeslong list, makes a heatmap
}
