% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_kmeans_cluster_hmap}
\alias{plot_kmeans_cluster_hmap}
\title{Plot Kmeans Clusters Heatmap}
\usage{
plot_kmeans_cluster_hmap(
  kl,
  col_order = NULL,
  logged = TRUE,
  base = NULL,
  pal = "RdBu",
  nrow = 2,
  col_fontsize = 6,
  row_fontsize = 6,
  legx = 0,
  legy = -0.45,
  labh = -1.1,
  labv = 1.1
)
}
\arguments{
\item{kl}{The kmeans clusters from `kmeans_by_selected_cols()`.}

\item{col_order}{Vector of column names in the order you want them to be drawn.}

\item{logged}{Indicates whether the fold changes are logged or not.}

\item{base}{If logged, the base that was used.}

\item{pal}{Palette to use from RColorBrewer.}

\item{nrow}{Number of rows to plot clusters in.}

\item{col_fontsize}{Font size for column names.}

\item{row_fontsize}{Font size for row names.}

\item{legx}{X position of legend in range (-0.5 .. 0.5).}

\item{legy}{Y position of legend in range(-0.5 .. 0.5).}

\item{labh}{Horizontal justification of subplot labels.}

\item{labv}{Vertical justification of subplot labels.}
}
\description{
This function plots the kmeans clusters as a heatmap.
}
\examples{
\dontrun{
# Example Usage:
plot_kmeans_cluster_hmap(my_kmeans_clusters,
col_order = c("Condition1", "Condition2"), logged = TRUE,
base = 2, pal = "RdBu", nrow = 2)
}
}
