% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_missing_peptides}
\alias{plot_missing_peptides}
\title{plot the representation of peptides in each group.}
\usage{
plot_missing_peptides(i)
}
\arguments{
\item{i}{dataframe with unmerged tech reps; typically from `import_data()`}
}
\value{
ggplot2 plot
}
\description{
Shows what proportion of the whole set of peptides has a NA value in each group
of treatment, seconds, bio rep and tech rep.
}
