% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_norm_qq}
\alias{plot_norm_qq}
\title{Create a Normal QQ plot of peptide quantity in biological replicates.}
\usage{
plot_norm_qq(i, log = FALSE, base = 2)
}
\arguments{
\item{i}{A data frame containing peptide quantity data.}

\item{log}{Logical, indicating whether to apply a logarithmic transformation to the quantity values. Default is FALSE.}

\item{base}{The logarithm base to use if 'log' is TRUE. Default is 2.}
}
\value{
A ggplot object representing the Normal QQ plot.
}
\description{
This function takes a data frame and generates a Normal QQ plot to visualize the distribution of peptide quantity in biological replicates. It can optionally apply a logarithmic transformation to the quantity values.
}
\examples{
\dontrun{
plot_norm_qq(data)
}

}
\seealso{
Other data visualization: 
\code{\link{plot_missing_peptides}()},
\code{\link{plot_pca}()},
\code{\link{plot_peptides_measured}()},
\code{\link{plot_quant_distributions}()},
\code{\link{plot_sample_kmeans}()}
}
\concept{data visualization}
\keyword{plot}
