% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_pca}
\alias{plot_pca}
\title{plots a pca on the treatment, seconds, bio-rep}
\usage{
plot_pca(df)
}
\arguments{
\item{df}{dataframe, typically from `import_data()`}
}
\value{
ggplot2 plot
}
\description{
Performs and draws a PCA plot with four panels, PCA with sample names
coloured by treatment, seconds and biorep and a scree plot of the PCA dimensions
}
