% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_quant_distributions}
\alias{plot_quant_distributions}
\title{draw density plots for data}
\usage{
plot_quant_distributions(i, log = FALSE, base = 2)
}
\arguments{
\item{log}{perform log transform of data}

\item{base}{base to use in log transform}

\item{df}{dataframe; typically from `import_data()`}
}
\value{
ggplot2 plot
}
\description{
Plot distribution of quantity for each treatment, seconds and biological
replicate. For biological replicate mean of technical replicates is taken
}
