% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_quant_distributions}
\alias{plot_quant_distributions}
\title{draw density plots for data}
\usage{
plot_quant_distributions(df, log = FALSE, base = 2)
}
\arguments{
\item{df}{dataframe; typically from `import_data()`}

\item{log}{perform log transform of data}

\item{base}{base to use in log transform}
}
\value{
ggplot2 plot
}
\description{
Plot density of quantities in data for each treatment, seconds and biological
replicate
}
