% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_vals}
\alias{replace_vals}
\title{replace missing values in a vector with others provided}
\usage{
replace_vals(x, lowest_vals)
}
\arguments{
\item{x}{vector of values with missing values}

\item{lowest_vals}{vector of replacements}
}
\value{
vector
}
\description{
replace missing values in a vector with others provided
}
