% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_columns_for_contrast}
\alias{select_columns_for_contrast}
\title{extract data columns for specified contrast}
\usage{
select_columns_for_contrast(
  l,
  treatment = NA,
  t_seconds = NA,
  control = NA,
  c_seconds = NA
)
}
\arguments{
\item{l}{list of data and row_info}

\item{treatment}{name of treatment to use}

\item{t_seconds}{value of seconds to use with treatment}

\item{control}{name of control to use}

\item{c_seconds}{value of seconds to use with control}
}
\value{
list with two members, `treatment` - matrix of treatment data; `control` -
matrix of control data
}
\description{
extract data columns for specified contrast
}
