% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{significant}
\alias{significant}
\title{Extract significant results}
\usage{
significant(x, alpha = NULL, by_fdr = TRUE)
}
\arguments{
\item{x}{A pepdiff_results object}

\item{alpha}{Significance threshold (default uses analysis alpha)}

\item{by_fdr}{Logical, use FDR-adjusted p-values (default TRUE)}
}
\value{
A tibble of significant results
}
\description{
Extract significant results
}
